requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.6
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
defusedxml==0.7.1
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.0.4
djangorestframework==3.13.1
et-xmlfile==1.1.0
kombu==5.2.4
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
packaging==21.3
Pillow==9.1.1
prompt-toolkit==3.0.29
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
